Refactor: Extract the COM event handler to a separate file
