+++ src/windows/webview_msedge.h
@@ -29,7 +29,6 @@
 
 namespace webview {
 using dispatch_fn_t = std::function<void()>;
-namespace detail {
 
 using msg_cb_t = std::function<void(const std::string)>;
 
@@ -165,11 +164,6 @@ get_last_native_path_component(const std::basic_string<T> &path) {
   return std::basic_string<T>();
 }
 #endif /* WEBVIEW_MSWEBVIEW2_BUILTIN_IMPL */
-
-template <typename T> struct cast_info_t {
-  using type = T;
-  IID iid;
-};
 
 namespace mswebview2 {
 static constexpr IID
@@ -418,190 +412,15 @@ private:
 #endif
 };
 
-namespace cast_info {
-static constexpr auto controller_completed =
-    cast_info_t<ICoreWebView2CreateCoreWebView2ControllerCompletedHandler>{
-        IID_ICoreWebView2CreateCoreWebView2ControllerCompletedHandler};
-
-static constexpr auto environment_completed =
-    cast_info_t<ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler>{
-        IID_ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler};
-
-static constexpr auto message_received =
-    cast_info_t<ICoreWebView2WebMessageReceivedEventHandler>{
-        IID_ICoreWebView2WebMessageReceivedEventHandler};
-
-static constexpr auto permission_requested =
-    cast_info_t<ICoreWebView2PermissionRequestedEventHandler>{
-        IID_ICoreWebView2PermissionRequestedEventHandler};
-} // namespace cast_info
 } // namespace mswebview2
 
-class webview2_com_handler
-    : public ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler,
-      public ICoreWebView2CreateCoreWebView2ControllerCompletedHandler,
-      public ICoreWebView2WebMessageReceivedEventHandler,
-      public ICoreWebView2PermissionRequestedEventHandler {
-  using webview2_com_handler_cb_t =
-      std::function<void(ICoreWebView2Controller *, ICoreWebView2 *webview)>;
+} // namespace webview
 
-public:
-  webview2_com_handler(HWND hwnd, msg_cb_t msgCb, webview2_com_handler_cb_t cb)
-      : m_window(hwnd), m_msgCb(msgCb), m_cb(cb) {}
+#include "webview2_loader/webview2_com_handler.hpp"
 
-  virtual ~webview2_com_handler() = default;
-  webview2_com_handler(const webview2_com_handler &other) = delete;
-  webview2_com_handler &operator=(const webview2_com_handler &other) = delete;
-  webview2_com_handler(webview2_com_handler &&other) = delete;
-  webview2_com_handler &operator=(webview2_com_handler &&other) = delete;
+using webview2_com_handler = webview::webview2_loader::webview2_com_handler;
 
-  ULONG STDMETHODCALLTYPE AddRef() { return ++m_ref_count; }
-  ULONG STDMETHODCALLTYPE Release() {
-    if (m_ref_count > 1) {
-      return --m_ref_count;
-    }
-    delete this;
-    return 0;
-  }
-  HRESULT STDMETHODCALLTYPE QueryInterface(REFIID riid, LPVOID *ppv) {
-    using namespace mswebview2::cast_info;
-
-    if (!ppv) {
-      return E_POINTER;
-    }
-
-    // All of the COM interfaces we implement should be added here regardless
-    // of whether they are required.
-    // This is just to be on the safe side in case the WebView2 Runtime ever
-    // requests a pointer to an interface we implement.
-    // The WebView2 Runtime must at the very least be able to get a pointer to
-    // ICoreWebView2CreateCoreWebView2EnvironmentCompletedHandler when we use
-    // our custom WebView2 loader implementation, and observations have shown
-    // that it is the only interface requested in this case. None have been
-    // observed to be requested when using the official WebView2 loader.
-
-    if (cast_if_equal_iid(riid, controller_completed, ppv) ||
-        cast_if_equal_iid(riid, environment_completed, ppv) ||
-        cast_if_equal_iid(riid, message_received, ppv) ||
-        cast_if_equal_iid(riid, permission_requested, ppv)) {
-      return S_OK;
-    }
-
-    return E_NOINTERFACE;
-  }
-  HRESULT STDMETHODCALLTYPE Invoke(HRESULT res, ICoreWebView2Environment *env) {
-    if (SUCCEEDED(res)) {
-      res = env->CreateCoreWebView2Controller(m_window, this);
-      if (SUCCEEDED(res)) {
-        return S_OK;
-      }
-    }
-    try_create_environment();
-    return S_OK;
-  }
-  HRESULT STDMETHODCALLTYPE Invoke(HRESULT res,
-                                   ICoreWebView2Controller *controller) {
-    if (FAILED(res)) {
-      // See try_create_environment() regarding
-      // HRESULT_FROM_WIN32(ERROR_INVALID_STATE).
-      // The result is E_ABORT if the parent window has been destroyed already.
-      switch (res) {
-      case HRESULT_FROM_WIN32(ERROR_INVALID_STATE):
-      case E_ABORT:
-        return S_OK;
-      }
-      try_create_environment();
-      return S_OK;
-    }
-
-    ICoreWebView2 *webview;
-    ::EventRegistrationToken token;
-    controller->get_CoreWebView2(&webview);
-    webview->add_WebMessageReceived(this, &token);
-    webview->add_PermissionRequested(this, &token);
-
-    m_cb(controller, webview);
-    return S_OK;
-  }
-  HRESULT STDMETHODCALLTYPE Invoke(
-      ICoreWebView2 *sender, ICoreWebView2WebMessageReceivedEventArgs *args) {
-    LPWSTR message;
-    args->TryGetWebMessageAsString(&message);
-    m_msgCb(webview::wstring::narrow_string(message));
-    sender->PostWebMessageAsString(message);
-
-    CoTaskMemFree(message);
-    return S_OK;
-  }
-  HRESULT STDMETHODCALLTYPE Invoke(
-      ICoreWebView2 *sender, ICoreWebView2PermissionRequestedEventArgs *args) {
-    COREWEBVIEW2_PERMISSION_KIND kind;
-    args->get_PermissionKind(&kind);
-    if (kind == COREWEBVIEW2_PERMISSION_KIND_CLIPBOARD_READ) {
-      args->put_State(COREWEBVIEW2_PERMISSION_STATE_ALLOW);
-    }
-    return S_OK;
-  }
-
-  // Checks whether the specified IID equals the IID of the specified type and
-  // if so casts the "this" pointer to T and returns it. Returns nullptr on
-  // mismatching IIDs.
-  // If ppv is specified then the pointer will also be assigned to *ppv.
-  template <typename T>
-  T *cast_if_equal_iid(REFIID riid, const cast_info_t<T> &info,
-                       LPVOID *ppv = nullptr) noexcept {
-    T *ptr = nullptr;
-    if (IsEqualIID(riid, info.iid)) {
-      ptr = static_cast<T *>(this);
-      ptr->AddRef();
-    }
-    if (ppv) {
-      *ppv = ptr;
-    }
-    return ptr;
-  }
-
-  // Set the function that will perform the initiating logic for creating
-  // the WebView2 environment.
-  void set_attempt_handler(std::function<HRESULT()> attempt_handler) noexcept {
-    m_attempt_handler = attempt_handler;
-  }
-
-  // Retry creating a WebView2 environment.
-  // The initiating logic for creating the environment is defined by the
-  // caller of set_attempt_handler().
-  void try_create_environment() noexcept {
-    // WebView creation fails with HRESULT_FROM_WIN32(ERROR_INVALID_STATE) if
-    // a running instance using the same user data folder exists, and the
-    // Environment objects have different EnvironmentOptions.
-    // Source: https://docs.microsoft.com/en-us/microsoft-edge/webview2/reference/win32/icorewebview2environment?view=webview2-1.0.1150.38
-    if (m_attempts < m_max_attempts) {
-      ++m_attempts;
-      auto res = m_attempt_handler();
-      if (SUCCEEDED(res)) {
-        return;
-      }
-      // Not entirely sure if this error code only applies to
-      // CreateCoreWebView2Controller so we check here as well.
-      if (res == HRESULT_FROM_WIN32(ERROR_INVALID_STATE)) {
-        return;
-      }
-      try_create_environment();
-      return;
-    }
-    // Give up.
-    m_cb(nullptr, nullptr);
-  }
-
-private:
-  HWND m_window;
-  msg_cb_t m_msgCb;
-  webview2_com_handler_cb_t m_cb;
-  std::atomic<ULONG> m_ref_count{1};
-  std::function<HRESULT()> m_attempt_handler;
-  unsigned int m_max_attempts = 5;
-  unsigned int m_attempts = 0;
-};
+namespace webview {
 
 class win32_edge_engine {
 public:
@@ -871,8 +690,6 @@ private:
   mswebview2::loader m_webview2_loader;
 };
 
-} // namespace detail
-
-using browser_engine = detail::win32_edge_engine;
+using browser_engine = win32_edge_engine;
 
 } // namespace webview
