+++ webview_test.cc
@@ -278,7 +278,7 @@ static void run_with_timeout(std::function<void()> fn, int timeout_ms) {
 // TEST: ensure that version number parsing works on Windows.
 // =================================================================
 static void test_parse_version() {
-  using namespace webview::detail;
+  using namespace webview;
   auto v = parse_version("");
   assert(v.size() == 4);
   assert(v[0] == 0 && v[1] == 0 && v[2] == 0 && v[3] == 0);
@@ -311,7 +311,7 @@ static void test_parse_version() {
 // TEST: ensure that narrow/wide string conversion works on Windows.
 // =================================================================
 static void test_win32_narrow_wide_string_conversion() {
-  using namespace webview::detail;
+//   using namespace webview;
   using namespace webview::wstring;
 
   assert(widen_string("").empty());
